% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_dt}
\alias{compute_dt}
\title{Compute time increments with optional clipping}
\usage{
compute_dt(t, clipped = TRUE)
}
\arguments{
\item{t}{Time vector (\eqn{n\_cells}).}

\item{clipped}{Logical. If \code{TRUE}, clip unusually large increments
using a Poisson-style upper bound. Default \code{TRUE}.}
}
\value{
Numeric vector of time increments.
}
\description{
Corresponds to scVelo:
\code{tools/_em_model_utils.py::compute_dt()}.
}
