% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_velocity.R
\name{.get_fit_time_matrix}
\alias{.get_fit_time_matrix}
\title{Retrieve and align fitted latent time}
\usage{
.get_fit_time_matrix(pv)
}
\arguments{
\item{pv}{A \code{plantvelo} object.}
}
\value{
A dense cells-by-genes fitted-time matrix, or \code{NULL}.
}
\description{
Retrieve and align fitted latent time
}
\keyword{internal}
