% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_velocity.R
\name{.velocity_axes}
\alias{.velocity_axes}
\title{Retrieve canonical velocity matrix axes}
\usage{
.velocity_axes(pv)
}
\arguments{
\item{pv}{A \code{plantvelo} object.}
}
\value{
A list containing canonical \code{cell_ids} and \code{gene_ids}.
}
\description{
Retrieve canonical velocity matrix axes
}
\keyword{internal}
