% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_dense_matrix}
\alias{make_dense_matrix}
\title{Convert to a dense numeric matrix}
\usage{
make_dense_matrix(X)
}
\arguments{
\item{X}{Sparse matrix, data frame, or ordinary matrix.}
}
\value{
Dense numeric matrix.
}
\description{
Converts a sparse matrix or data frame into a standard dense numeric matrix.
}
\details{
Corresponds to scVelo: \code{tools/utils.py::make_dense()}.
}
