% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null_coalesce}
\alias{null_coalesce}
\alias{\%||\%}
\title{NULL-coalescing operator}
\usage{
a \%||\% b
}
\arguments{
\item{a}{Primary value.}

\item{b}{Fallback value.}
}
\value{
Either \code{a} or \code{b}.
}
\description{
Returns \code{a} if \code{a} is not \code{NULL}; otherwise returns
\code{b}.
}
\keyword{internal}
