% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity_plot.R
\name{plot_latent_time}
\alias{plot_latent_time}
\title{Plot Latent Time}
\usage{
plot_latent_time(pv, reduction = "umap", ...)
}
\arguments{
\item{pv}{A \code{plantvelo} object.}

\item{reduction}{Character scalar. Embedding to plot. Default
\code{"umap"}.}

\item{...}{Further arguments passed to \code{plot_velocity_embedding()}.}
}
\value{
A \code{ggplot} object.
}
\description{
Convenience wrapper around \code{plot_velocity_embedding()} that colours
cells by \code{latent_time} and disables velocity arrows.
}
\examples{
\dontrun{
plot_latent_time(pv)
}
}
