% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tau_inv_u}
\alias{tau_inv_u}
\title{Infer relative time \eqn{\tau} from U values}
\usage{
tau_inv_u(u, u0, alpha, kappa)
}
\arguments{
\item{u}{Observed \eqn{U} values.}

\item{u0}{Initial amount of \eqn{U}.}

\item{alpha}{Transcription rate into \eqn{U}.}

\item{kappa}{Total consumption rate of \eqn{U}.}
}
\value{
Numeric vector of inferred times \eqn{\tau}.
}
\description{
Given observed \eqn{U} values, inverts the analytical solution for
\eqn{U(t)} to recover the corresponding relative time \eqn{\tau}.
}
\details{
Corresponds to scVelo:
\code{tools/_em_model_utils.py::tau_inv()} (U branch).
}
